backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.6.0
PyJWT==2.10.1
bcrypt==5.0.0
channels==4.3.2
python-dotenv==1.2.1
